Fix #31: When selecting a Glyph in the GlyphsView the view scrolls to the start.
GlyphProperties was including font change since the glyph metrics needed to be updated when the font properties change; however, this was causing erroneous FontFamily and FontSize changed event.

The change does the following:
* Add a specific GlyphMetrics flag.
* Explicitly set this flag if font, fontsize or glyph changes
* Update GlypProperties to include Glyph and GlyphMetrics flag.

UpdateFont is still called if any font flags are set and UpdateGlyph is called when any glyph flags are set.

GlyphChangedEvent and GlyphMetricsChangedEvent are now tied to the explicit flag.